+++ MediaBrowser.Model/Configuration/MetadataPluginSummary.cs
@@ -2,7 +2,7 @@
 #pragma warning disable CS1591
 
 using System;
-using MediaBrowser.Model.Entities;
+using Jellyfin.Data.Enums;
 
 namespace MediaBrowser.Model.Configuration
 {
